requirements.txt
streamlit
Pillow
pandas
pdfkit
jinja2